Remove separate GitRelease files.
