+++ tubular/github_api.py
@@ -1,8 +1,10 @@
 """ Provides Access to the GitHub API """
 from __future__ import print_function, unicode_literals
 
-import os
+from datetime import datetime, timedelta
 import logging
+import os
+import string
 
 from github import Github
 from github.Commit import Commit
@@ -13,11 +15,45 @@ LOGGER = logging.getLogger(__name__)
 LOGGER.setLevel(logging.INFO)
 
 
+# Day of week constant
+_TUESDAY = 1
+_NORMAL_RELEASE_WEEKDAY = _TUESDAY
+
+
 class NoValidCommitsError(Exception):
     """
     Error indicating that there are no commits with valid statuses
     """
     pass
+
+
+def extract_message_summary(message, max_length=50):
+    """
+    Take a commit message and return the first part of it.
+    """
+    title = string.split(message, '\n')[0] or ''
+    if len(title) < max_length:
+        return title
+    else:
+        return title[0:max_length] + '...'
+
+
+def default_expected_release_date(release_day=_NORMAL_RELEASE_WEEKDAY):
+    """
+    Returns the default expected release date given the current date.
+    Currently the nearest Tuesday in the future (can't be today)
+    """
+    proposal = datetime.now() + timedelta(days=1)
+    while proposal.weekday() is not release_day:
+        proposal = proposal + timedelta(days=1)
+    return proposal
+
+
+def rc_branch_name_for_date(date):
+    """
+    Returns the standard release candidate branch name
+    """
+    return 'rc/{date}'.format(date=date.isoformat())
 
 
 class GitHubAPI(object):
